# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Web App API
  description: |
    REST API for Movie Web App.

    ## Authentication
    Using **Google OAuth2** for login and **JWT Bearer Token** for accessing endpoints that require authentication.

    ## Roles
    - **guest** — Public access without authentication
    - **user** — User that has logged in via Google OAuth
    - **admin** — Administrator with full access
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://movie-ku-api.vercel.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from Google OAuth login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
      required: [success, message]

    # --- Auth ---
...
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required: [id, name]

    # --- Movie Image ---
    MovieImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image_type:
          type: string
          enum: [poster, backdrop]
        image_url:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true

    # --- Movie Video ---
    MovieVideo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        video_type:
          type: string
          enum: [trailer, teaser]
        site:
          type: string
          enum: [youtube, vimeo]
        video_key:
          type: string
        official:
          type: boolean

    # --- Movie ---
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
        api_id:
          type: string
          nullable: true
          description: TMDB API ID (null if user-created)
        source:
          type: string
          enum: [tmdb, user]
        title:
          type: string
        overview:
          type: string
          nullable: true
        release_date:
          type: string
          format: date
          nullable: true
        popularity:
          type: number
          format: float
          nullable: true
        rating:
          type: number
          format: float
          nullable: true
        is_featured:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        created_by:
          type: string
          format: uuid
          nullable: true
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        images:
          type: array
          items:
            $ref: "#/components/schemas/MovieImage"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/MovieVideo"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, source, title, status, created_at, updated_at]
    MovieCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        overview:
          type: string
        release_date:
          type: string
          format: date
        genre_ids:
          type: array
          items:
            type: string
            format: uuid
          description: List of genre UUIDs
      required: [title]
    MovieUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        overview:
          type: string
        release_date:
          type: string
          format: date
        genre_ids:
          type: array
          items:
            type: string
            format: uuid
    AdminMovieCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        overview:
          type: string
        release_date:
          type: string
          format: date
        popularity:
          type: number
          format: float
        rating:
          type: number
          format: float
        is_featured:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        genre_ids:
          type: array
          items:
            type: string
            format: uuid
      required: [title]
    AdminMovieUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        overview:
          type: string
        release_date:
          type: string
          format: date
        popularity:
          type: number
          format: float
        rating:
          type: number
          format: float
        is_featured:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        genre_ids:
          type: array
          items:
            type: string
            format: uuid

    # --- Wishlist ---
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
        total_pages:
          type: integer
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        enum: [title, release_date, popularity, rating, created_at]
        default: created_at
      description: Field to sort by
    OrderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
paths:
...
  /api/movies:
    get:
      tags: [Movies]
      summary: List all movies
      operationId: listMovies
      description: |
        Browse movies with support for search, filtering, sorting, and pagination.
        Accessible by anyone without authentication.
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by movie title (partial match)
        - name: genre_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by genre UUID
        - name: source
          in: query
          schema:
            type: string
            enum: [tmdb, user, admin]
          description: Filter by movie source
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
            default: active
          description: Filter by status
        - name: release_date_from
          in: query
          schema:
            type: string
            format: date
          description: Filter movies released on or after this date
        - name: release_date_to
          in: query
          schema:
            type: string
            format: date
          description: Filter movies released on or before this date
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/OrderParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: List of movies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Movie"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
  /api/movies/popular:
    get:
      tags: [Movies]
      summary: Get popular movies
      operationId: getPopularMovies
      description: Returns movies sorted by popularity (descending).
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: Popular movies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Movie"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
  /api/movies/{id}:
    get:
      tags: [Movies]
      summary: Get movie detail
      operationId: getMovieDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Movie detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== USER MOVIES ====================
  /api/movies/me:
    get:
      tags: [User Movies]
      summary: List current user's movies
      operationId: listMyMovies
      description: Returns all movies created by the authenticated user.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: User's movies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Movie"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/movies/user:
    post:
      tags: [User Movies]
      summary: Create a movie
      operationId: createUserMovie
      description: Create a new movie entry. The movie will be tagged with `source=user` and `created_by` set to the current user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieCreateRequest"
      responses:
        "201":
          description: Movie created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/movies/user/{id}:
    put:
      tags: [User Movies]
      summary: Update own movie
      operationId: updateUserMovie
      description: Update a movie owned by the current user. Cannot update movies from TMDB or other users.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieUpdateRequest"
      responses:
        "200":
          description: Movie updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not the movie owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [User Movies]
      summary: Delete own movie
      operationId: deleteUserMovie
      description: Soft-delete a movie owned by the current user.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Movie deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not the movie owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== WISHLIST ====================
...
  /api/admin/movies:
    get:
      tags: [Admin Movies]
      summary: List all movies (admin)
      operationId: listAdminMovies
      description: List all movies including archived. Admin can see all sources.
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [tmdb, user, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/OrderParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: All movies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Movie"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Admin Movies]
      summary: Create movie (admin)
      operationId: createAdminMovie
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMovieCreateRequest"
      responses:
        "201":
          description: Movie created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/movies/{id}:
    put:
      tags: [Admin Movies]
      summary: Update movie (admin)
      operationId: updateAdminMovie
      description: Admin can update any movie regardless of source or ownership.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMovieUpdateRequest"
      responses:
        "200":
          description: Movie updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Admin Movies]
      summary: Delete movie (admin)
      operationId: deleteAdminMovie
      description: Soft-delete any movie. Admin can delete movies from any source.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Movie deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== ADMIN SYNC ====================
  /api/admin/sync/movies:
    post:
      tags: [Admin Sync]
      summary: Trigger movie sync from TMDB
      operationId: triggerMoviesSync
      description: |
        Start a background sync with TMDB API. Supports two modes:
        - `full` (default): Sync all movies from popular + now_playing endpoints
        - `changes`: Incremental sync — only movies changed in the last 14 days

        The sync runs asynchronously. Use GET /sync/status to monitor progress.
        Returns 409 if a sync is already running.
      security:
        - BearerAuth: []
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [full, changes]
            default: full
          description: |
            Sync mode. `full` syncs all available pages from TMDB list endpoints.
            `changes` uses TMDB /movie/changes API (last 14 days).
        - name: resume
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
          description: |
            Resume from last failed sync position. Only applies to `full` mode.
      responses:
        "202":
          description: Sync started in background
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Movies sync started in background"
                          type:
                            type: string
                            enum: [movies, changes]
                          started_at:
                            type: string
                            format: date-time
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A sync is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...